fix update error

+++ ApiServer/DsFinal/src/main/java/com/doubleslash/fifth/vo/UserVO.java
@@ -28,8 +28,8 @@ public class UserVO {
 	
 	private String nickname;
 	
-	private double drink;
+	private float drink;
 	
-	private Integer hangover;
+	private int hangover;
 	
 }
